requirements.txt
feedparser==6.0.10
google-generativeai==0.3.2
requests==2.31.0
tqdm==4.66.1
python-dotenv==1.0.0
flask==2.3.3 